chore: adapt annotations to stricter numba dispatcher type stubs
The bumped numba ships Dispatcher stubs that type-check calls against the wrapped function's annotations: float64 constants are now typed as such, rand_int32(_array) accept plain ints alongside np.int32, and a numpy bool comparison result is converted before being passed as a bool parameter.

+++ src/max_div/_core/_math/fast_log_exp/_fast_exp.py
@@ -10,7 +10,7 @@ from max_div._core._math.powers_of_2 import power_of_2_f32
 # -------------------------------------------------------------------------
 
 # --- float64 ---------------------------------------------
-_D_LOG2_E = 1.4426950408889634  # np.log2(e)
+_D_LOG2_E = np.float64(1.4426950408889634)  # np.log2(e)
 
 # Obtained via minimax polynomial fitting over [0.0, 1.0) with additional
 # continuity & smoothness constraints imposed on boundary points.
